+++ server/areanode.h
@@ -26,7 +26,7 @@ typedef struct areanode_s
 	link_t		portal_edicts;
 } areanode_t;
 
-#define STRUCT_FROM_LINK( l, t, m )		((t *)((byte *)l - (int)&(((t *)0)->m)))
+#define STRUCT_FROM_LINK( l, t, m )		((t *)((byte *)l - (size_t)&(((t *)0)->m)))
 #define EDICT_FROM_AREA( l )			STRUCT_FROM_LINK( l, edict_t, area )
 #define FACET_FROM_AREA( l )			STRUCT_FROM_LINK( l, mfacet_t, area )
